refactor: shop table row 컴포넌트 한 파일로 정리
