+++ apps/web/src/app/[locale]/shop/SellListSection/index.tsx
@@ -3,39 +3,26 @@
 import { useState } from 'react';
 import { useTranslations } from 'next-intl';
 import type { Product } from '@gitanimals/api';
-import useIsMobile from '@gitanimals/react/src/hooks/useIsMobile/useIsMobile';
 import { auctionQueries } from '@gitanimals/react-query';
 import { Button } from '@gitanimals/ui-panda';
 import { useQuery } from '@tanstack/react-query';
 
+import { MediaQuery } from '@/components/MediaQuery';
 import Pagination from '@/components/Pagination/Pagination';
-import { ShopTableMobileRow } from '@/components/ProductTable/ShopTableMobileRow';
-import ShopTableRowView, { ShopTableRowViewSkeleton } from '@/components/ProductTable/ShopTableRowView';
-import { ACTION_BUTTON_OBJ } from '@/constants/action';
 
+import { ShopTableDesktopRow, ShopTableMobileRow, ShopTableRowViewSkeleton } from '../_common/ShopTableMobileRow';
 import { tableCss, tbodyCss, theadCss } from '../AuctionSection/table.styles';
 
 import EditModal from './EditModal';
 
-const SELL_LIST_ACTION_OBJ = ACTION_BUTTON_OBJ['EDIT'];
-
 function SellListSection() {
   const t = useTranslations('Shop');
-  const isMobile = useIsMobile();
 
   const [editProductId, setEditProductId] = useState<string>();
   const [currentPage, setCurrentPage] = useState(0);
 
   const { data } = useQuery({
-    ...auctionQueries.myProductsOptions({
-      pageNumber: currentPage,
-    }),
-    select: (data) => ({
-      ...data,
-      products: data.products.map((product) => {
-        return { ...product };
-      }),
-    }),
+    ...auctionQueries.myProductsOptions({ pageNumber: currentPage }),
     placeholderData: (prevData) => prevData,
   });
 
@@ -54,42 +41,37 @@ function SellListSection() {
           <span>{t('price')}</span>
           <span></span>
         </div>
-
         <div className={tbodyCss}>
           {!data && Array.from({ length: 8 }).map((_, index) => <ShopTableRowViewSkeleton key={`skeleton-${index}`} />)}
-          {data?.products.map((product) => {
-            return isMobile ? (
-              <ShopTableMobileRow
-                key={product.id}
-                personaType={product.persona.personaType}
-                personaLevel={product.persona.personaLevel}
-                price={product.price}
-                rightElement={
-                  <Button
-                    size="s"
-                    variant="secondary"
-                    onClick={() => onEditAction(product.id)}
-                    color={SELL_LIST_ACTION_OBJ.color}
-                  >
-                    {t('edit')}
-                  </Button>
-                }
-              />
-            ) : (
-              <ShopTableRowView
-                key={product.id}
-                id={product.id}
-                persona={product.persona}
-                price={product.price}
-                rightElement={
-                  <Button variant="secondary" onClick={() => onEditAction(product.id)}>
-                    {' '}
-                    {t('edit')}
-                  </Button>
-                }
-              />
-            );
-          })}
+          {data?.products.map((product) => (
+            <MediaQuery
+              key={product.id}
+              mobile={
+                <ShopTableMobileRow
+                  personaType={product.persona.personaType}
+                  personaLevel={product.persona.personaLevel}
+                  price={product.price}
+                  rightElement={
+                    <Button variant="secondary" size="s" onClick={() => onEditAction(product.id)}>
+                      {t('edit')}
+                    </Button>
+                  }
+                />
+              }
+              desktop={
+                <ShopTableDesktopRow
+                  personaType={product.persona.personaType}
+                  personaLevel={product.persona.personaLevel}
+                  price={product.price}
+                  rightElement={
+                    <Button variant="secondary" onClick={() => onEditAction(product.id)}>
+                      {t('edit')}
+                    </Button>
+                  }
+                />
+              }
+            />
+          ))}
         </div>
       </div>
